Module advent2020::day04

Day 4: Interestingly, in contrast to Day 3, we achieve better performance by mapping before folding in this case. This is another good example of where splitn outperforms split, even if some strings have fewer than n chunks. It seems that parsing the passport is complex enough that combining is not as important.

Structs

PassportData

PassportValidityData

Functions

byr_valid

ecl_valid

eyr_valid

hcl_valid

hgt_valid

iyr_valid

part_1

part_2

pid_valid

report

run