Module advent2020::day14

Day 14: The bitwise operations are straightforward, but the combinatorics make this tricky. 64 bit integers are slower to work with, and my current solution to Part 2 is slow in hitting all required floating mask values. I sum the data as I walk the list, visiting the instructions in reverse, to avoid repetition in traversing the map.

Structs

Instructions

Update

Constants

CAPACITY

INSTRUCTIONS

MAX_36_BITS

Functions

report

run

update_memory_1

update_memory_2