Module advent2020::day17

Day 17: I have experimented with maintaining a hash set of only the active nodes, but this turned into an interesting example where doing lots of the same task is faster than some special case logic to reduce the total FLOPs. In the 4D case, the final solution uses ~1% of the cells, but trying to target active cells and neighbors took 10x longer on my machine. I did use symmetry in the 3rd and 4th dimension to reduce the computation by a factor of nearly 2 and 4, respectively.

Constants

CAPACITY_3D

CAPACITY_4D

CYCLES

NEIGHBOR_LIMIT

Functions

count_neighbors_3d

count_neighbors_4d

game_of_life_3d

game_of_life_4d

report

run